fix for #7
additional fixes for the time_marker switch to double

+++ src/main/java/com/thundermoose/bio/model/NormalizedData.java
@@ -11,7 +11,7 @@ public class NormalizedData {
 
   }
 
-  public NormalizedData(String plateName, String geneId, int timeMarker, float normalized) {
+  public NormalizedData(String plateName, String geneId, double timeMarker, float normalized) {
     this.plateName = plateName;
     this.geneId = geneId;
     this.timeMarker = timeMarker;
